requirements.txt
scikit-image
pillow
networkx
nltk
tensorflow
keras-cv
keras-nlp
keras-tuner
tf_keras
numpy
scipy
torch
torchvision
openai-clip
google-genai
OpenAI
anthropic
spacy
